fixed float issue

+++ Interpreter/Values/Float.cs
@@ -16,12 +16,12 @@ public class Float : BaseValue, IValue
 
     public override BaseValue OperatedBy(Token _operator, BaseValue other)
     {
-        if (Value is not double && other.Value is not double)
+        if (Value is not float && other.Value is not float)
         {
-            throw new TypeConversionException(typeof(double));
+            throw new TypeConversionException(typeof(float));
         }
-        var value = (double)Value;
-        var otherValue = (double)other.Value;
+        var value = (float)Value;
+        var otherValue = (float)other.Value;
         switch (_operator.TokenType)
         {
             case TokenOperators.PLUS:
